+++ lithops/serverless/backends/cloudrun/__init__.py
@@ -1,1 +1,0 @@
-from .cloudrun import CloudRunServingBackend as ServerlessBackend